scientific-skills/biopython/scripts/requirements.txt
# Biopython - Biological Computation
biopython>=1.81
numpy>=1.21.0
requests>=2.25.0
